Refactored SnapshotMap component
